+++ docs/changelog.rst
@@ -56,6 +56,8 @@ Fixes:
 * Fix a bug, where the autotagger still considers matches that are specifically
   listed under the config's ignored section. :bug:`1487`
 * Fix a bug with unicode strings when generating thumbnails. :bug:`1485`
+* :doc:`/plugins/plexupdate`: Fixed library updates not working when in a Plex
+  Home by allowing a token with requests.
 
 
 1.3.13 (April 24, 2015)